Flex: reuse Box story's argTypes configuration
